% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lusweave-package.R
\docType{package}
\name{lusweave-package}
\alias{lusweave}
\alias{lusweave-package}
\title{lusweave: Sweave/Knitr Utilities}
\description{
Set of tools which simplify the usage of SWeave/Knitr in R and allow to easily create PDF files from within R.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/pik-piam/lusweave}
  \item \doi{10.5281/zenodo.1158594}
  \item Report bugs at \url{https://github.com/pik-piam/lusweave/issues}
}

}
\author{
\strong{Maintainer}: Jan Philipp Dietrich \email{dietrich@pik-potsdam.de}

Authors:
\itemize{
  \item Markus Bonsch
  \item David Klein \email{dklein@pik-potsdam.de}
  \item Florian Humpenoeder \email{humpenoeder@pik-potsdam.de}
}

}
\keyword{internal}
