% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lusweavedata.R
\name{lusweavedata}
\alias{lusweavedata}
\title{lusweavedata}
\description{
General lusweave-dataset
}
\details{
Please do not directly access that data. It should be only used by library functions.
}
\author{
Jan Philipp Dietrich
}
