% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/swclose.R
\name{swclose}
\alias{swclose}
\title{Close a swStream}
\usage{
swclose(
  stream,
  outfile = "",
  latexpath = "",
  clean_output = TRUE,
  engine = "knitr",
  save_stream = FALSE,
  knitquiet = TRUE
)
}
\arguments{
\item{stream}{The \code{swStream} object to be processed.}

\item{outfile}{Name of the pdf to be produced. Overwrites the name given in
swopen}

\item{latexpath}{Path of the LaTeX distribution (only necessary if not part
of the PATH variable).}

\item{clean_output}{If true, all auxiliary files will be deleted. If False,
they remain, including each plot in its own pdf.}

\item{engine}{Engine to use for conversion. Currently available: Sweave and
knitr.}

\item{save_stream}{If true (default is false) stream is saved to .rda file.}

\item{knitquiet}{If false (default) progressbar and messages are printed
otherwise suppressed.}
}
\value{
No return value.
}
\description{
Function to create a pdf from a \code{"\linkS4class{swStream}"} object
}
\details{
Creates a pdf with of the name specified in \code{stream@name}, with the
content specified in \code{stream@content}, using \code{stream@Sweave.sty}
as the style file.
}
\examples{
\dontrun{
test <- swopen(outfile = "test.pdf")
swlatex(test, "tttteeeesssstttt")
swclose(test)
# Change the name
swclose(test, outfile = "test_2.pdf")
}

}
\seealso{
\code{"\linkS4class{swStream}"},\code{\link{swopen}},\code{\link{swlatex}},\code{\link{swR}},\code{\link{swtable}},\code{\link{swfigure}}
}
\author{
Markus Bonsch, David Klein
}
