% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/swopen.R
\name{swopen}
\alias{swopen}
\title{Open an swStream}
\usage{
swopen(
  outfile = "out.pdf",
  folder = "",
  template = "default",
  style = "default",
  orientation = NULL,
  envir = new.env()
)
}
\arguments{
\item{outfile}{name for the output pdf. Can be specified with or w/o the
.pdf extension. Can also contain a path to a differetn directory.}

\item{folder}{Path where the output shall be produced.}

\item{template}{A template which should be used for the PDF. Either a path
to a .tex-file or a vector containing the tex template code or a name of a
template which already comes with the library. Currently the following
templates are part of the library: "default","default_landscape" and
"david".}

\item{style}{Style information which should be used for the PDF. Either a
path to a .sty-file or a vector containing the tex style code or a name of a
style which already comes with the library. Currently the following styles
are part of the library: "default"}

\item{orientation}{[DEPRECATED] Please do not use this argument. It is just
left in the function for compatibility reasons.}

\item{envir}{The environment in which the object should be saved. A new
environment by default.}
}
\value{
An environment containing a \code{"\linkS4class{swStream}"} object
named "stream" with a header for the Rnw file and the content of a
stylefile.
}
\description{
Function to create a swStream object, that can be used to create a pdf
including R output
}
\examples{
\dontrun{
test<-swopen()
str(test)

testtwo<-swopen(outfile="test.pdf")
str(testtwo)

}
}
\seealso{
\code{"\linkS4class{swStream}"},\code{\link{swclose}},\code{\link{swlatex}},\code{\link{swR}},\code{\link{swtable}},\code{\link{swfigure}}
}
\author{
Markus Bonsch, Jan Philipp Dietrich
}
