% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/swStream-class.R
\docType{class}
\name{swStream-class}
\alias{swStream-class}
\title{Class "swStream" ~~~}
\description{
The swStream class provides an interface to the Sweave method and allows for
easy creation of pdf's from R output.
}
\details{
The Sweave method is an interface between R and LaTeX. It processes .Rnw
files, that can contain LaTeX code as well as R code into proper .tex files,
replacing the R commands by their output. See
\code{http://www.stat.uni-muenchen.de/~leisch/Sweave/} for details.\cr An
additional style file is needed to compile the .tex file into a pdf.
}
\section{Objects from the Class}{
 Objects can be created by calls of the form
\code{new("swStream", name,folder, content, Sweave.sty, arguments,
functions, auxfiles, envir)}.
}

\examples{
\dontrun{
showClass("swStream")
}
}
\seealso{
\code{\link{swopen}},\code{\link{swclose}},\code{\link{swlatex}},\code{\link{swR}},\code{\link{swtable}},\code{\link{swfigure}}
}
\author{
Markus Bonsch
}
\keyword{classes}
