% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/swtable.R
\name{swtable}
\alias{swtable}
\title{swtable}
\usage{
swtable(
  stream,
  x,
  caption = NULL,
  label = NULL,
  transpose = FALSE,
  digits = NULL,
  vert.lines = 1,
  hor.lines = 1,
  align = "c",
  display = NULL,
  colsplit = NULL,
  ...
)
}
\arguments{
\item{stream}{The \code{swStream} object to be modified.}

\item{x}{An array or something that can be converted to an array, that shall
be shown as a table. X should only contain 2 dimensions with more than 1
element, otherwise it cannot be plotted as a matrix.}

\item{caption}{Description to be displayed in the pdf.}

\item{label}{Label for the tex file for referencing.}

\item{transpose}{Should the matrix be transposed before it is plotted?}

\item{digits}{How many digits shall be shown for each number? Either one
number or ncol()+1 numbers for each column seperately.}

\item{vert.lines}{1 for vertical lines between the columns, 0 for no
vertical lines. If you want lines only somewhere, set this to 0 and use
align.}

\item{hor.lines}{1 for horizontal lines between the rows, 0 for no
horizontal lines. A vector of length equal to nrow(x)+2 (for line above
first row also) with 1 and 0 indicates to put lines wherever a one is
mentioned, ohters will stay without lines. A vector of numbers between -1
and "nrow(x)", inclusive, indicating the rows after which a horizontal line
should appear.}

\item{align}{Alignment of the table content. You can specify a single
character: "c" for centered, "l" for left, "r" for right. If you want
different alignments for different columns, specify a vector of length equal
to the table columns plus one (for the row names) with the alignment for
each column. If you want vertical lines only between some columns or around
the table, set vert.lines to 0 and add "|" to the alignment vector, where
you want to have vertical lines.}

\item{display}{How shall the content be displayed? Use "d" (for integers),
"f", "e", "E", "g", "G", "fg" (for reals), or "s" (for strings).}

\item{colsplit}{Number of columns after which the table is split into two (Integer).
Table will not be splitted up if set to NULL.}

\item{...}{Further options passed to print.xtable.}
}
\value{
No return value.
}
\description{
Function to add an R object to a \code{"\linkS4class{swStream}"} object.
}
\details{
This method provides the possibility to display the content of an array-like
object in a table. It has to be specified, which two dimensions should be
shown by choosing them before the object is sent to the function.  The
method is built on the \code{xtable} method. The table is then created with
the \code{print.xtable} method, that takes all kind of arguments like
placement etc. See that manuals for more information
}
\examples{
\dontrun{
sw <- swopen(outfile = "test.pdf")
a <- array(1:45, c(5, 2, 3), dimnames = list(paste0("a", 1:5), paste0("b", 1:2), paste0("c", 1:3)))
print(a)
swtable(sw, a[, , 1], "test 1")
swtable(sw, a[, 1, ], "test 2")
swtable(sw, a[, 1, ], "transpose test", transpose = TRUE)
swtable(sw, a[, 1, ], "colsplit test", transpose = TRUE, colsplit = 2)
swclose(sw)
}
}
\seealso{
\code{"\linkS4class{swStream}"},\code{\link{swopen}},\code{\link{swclose}},\code{\link{swlatex}},\code{\link{swR}},\code{\link{swfigure}}
}
\author{
Markus Bonsch, Jan Philipp Dietrich
}
